Cleaned up App.js, moved SwipableNav to new folder, added Template and TemplateContainer to begin site buildout.
