+++ src/components/TemplateContainer.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Template from './Template';
+
+class TemplateContainer extends React.Component {
+
+  render() {
+    return (
+      <Template />
+    );
+  }
+}
+
+export default TemplateContainer
